move exec metrics to own file
